+++ Docs/ReleaseNotes.md
@@ -24,6 +24,7 @@ For breaking API changes see [this document](https://github.com/jrouwe/JoltPhysi
 * Added support for glTF `KHR_physics_rigid_bodies` constraint motors. This adds `ESpringMode::MassNormalizedStiffnessAndDamping` to be able to specify the spring parameters mass-normalized / in acceleration-mode. It also adds `EMotorState::PositionAndVelocity` which creates a force proportional to `stiffness * (position_target - position_current) + damping * (velocity_target - velocity_current)`.
 * Added `Ragdoll::DriveToPoseUsingMotors` variant that drives to a pose using both position and velocity.
 * Added `Body::ApplyBodyCreationSettings` and `Body::ApplySoftBodyCreationSettings` to be able to update a body with creation settings after creation.
+* Added `ShapeCastSettings::mExtraConvexRadius` which inflates the query shape by an extra convex radius.
 * Various performance and memory optimizations.
 
 ### Bug Fixes